@@ -2,18 +2,26 @@
   "name": "short-lnk",
   "private": true,
   "scripts": {
-    "start": "meteor run"
+    "start": "meteor run",
+    "test": "meteor test --driver-package=practicalmeteor:mocha"
   },
   "dependencies": {
-    "babel-runtime": "6.18.0",
+    "babel-runtime": "^6.18.0",
     "meteor-node-stubs": "~0.2.0",
     "moment": "^2.17.1",
-    "react": "^15.4.1",
-    "react-dom": "^15.4.1",
-    "react-router": "^3.0.0",
-    "simpl-schema": "0.0.3"
+    "react": "^15.6.2",
+    "react-addons-pure-render-mixin": "^15.4.1",
+    "react-dom": "^15.6.2",
+    "react-router": "^3.2.1",
+    "simpl-schema": "0.0.3",
+    "simple-schema": "^1.1.0"
   },
   "engines": {
     "node": "4.6.2"
+  },
+  "devDependencies": {
+    "enzyme": "^2.7.1",
+    "expect": "^1.20.2",
+    "react-addons-test-utils": "^15.4.1"
   }
 }
