@@ -2,7 +2,7 @@ import { Meteor } from 'meteor/meteor';
 import SimpleSchema from 'simpl-schema';
 import { Accounts } from 'meteor/accounts-base';
 
-Accounts.validateNewUser((user) => {
+export const validateNewUser = (user) => {
   const email = user.emails[0].address;
 
   new SimpleSchema({
@@ -13,4 +13,8 @@ Accounts.validateNewUser((user) => {
   }).validate({ email });
 
   return true;
-});
+};
+
+if(Meteor.isServer) {
+  Accounts.validateNewUser(validateNewUser);
+}
