@@ -1,8 +1,9 @@
 import React from 'react';
 import { Link } from 'react-router';
 import { Accounts } from 'meteor/accounts-base';
+import { createContainer } from 'meteor/react-meteor-data';
 
-export default class Signup extends React.Component {
+export class Signup extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
@@ -19,7 +20,7 @@ export default class Signup extends React.Component {
       return this.setState({error: 'Password must be more than 8 characters long'});
     }
 
-    Accounts.createUser({email, password}, (err) => {
+    this.props.createUser({email, password}, (err) => {
       if (err) {
         this.setState({error: err.reason});
       } else {
@@ -47,3 +48,13 @@ export default class Signup extends React.Component {
     );
   }
 }
+
+Signup.propTypes = {
+  createUser: React.PropTypes.func.isRequired
+}
+
+export default createContainer(() => {
+  return {
+    createUser: Accounts.createUser
+  }
+}, Signup);
